backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
